==> Notebooks/9.GraphViz.ipynb <==
# Graph Visualization with NetworkX/PyGraphistry

from __future__ import print_function
import graphistry
import snap
from snap import TUNGraph
import pandas as pd
import matplotlib as mpl
from matplotlib import pyplot as plt
import seaborn as sns
import os
import numpy as np
import networkx as nx

# Set Seaborn defaults
sns.set()
%matplotlib inline
pd.set_option("display.precision", 6)
mpl.rcParams['figure.dpi'] = 100
mpl.rcParams['savefig.dpi'] = 100
mpl.rcParams['figure.autolayout'] = True

# Global variables
data_dir = "../data"
pictures_path = os.path.join("../Pictures", "9.GraphViz")

def load_graph_from_backup(filename):
    FIn = snap.TFIn("../data/"+filename+".bin")
    graph = snap.TUNGraph.Load(FIn)
    return graph

hashtag = "jesuischarlie"
hashtag_subgraph = load_graph_from_backup("mmr_subgraph_"+hashtag)

# Convert from SNAP format to Pandas DF
graph_df = pd.DataFrame(columns=["src","dst"], dtype=np.uint32)
for e in hashtag_subgraph.Edges():
    temp = pd.DataFrame([[e.GetSrcNId(), e.GetDstNId()]], columns=["src","dst"], dtype=np.uint32)
    graph_df = graph_df.append(temp)

## 1. Load graph data from Pandas DataFrames

hashtag = "jesuischarlie"

edges_df = pd.read_csv(os.path.join(data_dir,"graph_visualization_top_interactions/"+hashtag+"_top_interactions.csv"), dtype={"Source": np.object, "Target": np.object, "Count": np.uint32})
nodes_df = pd.read_csv(os.path.join(data_dir,"graph_visualization_top_interactions/"+hashtag+"_users_frequency.csv"), dtype={"User": np.object, "Target": np.object, "Count": np.uint32})

# Build dict of node attrs
nodes_attr = {}
for _,r in nodes_df.iterrows():
    nodes_attr[r["User"]] = r["Count"]
nodes_attr = {k:nodes_attr[k] for k in nodes_attr if G.has_node(k)}

# Build graph from list of edges
G = nx.Graph()
G.add_weighted_edges_from(edges_df.values)

# Set node attributes
nx.set_node_attributes(G, nodes_attr, name="degree")

# Plot graph
